requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
groq==0.9.0
python-dotenv==1.0.1
openai==1.35.10
streamlit==1.36.0
streamlit-carousel==1.0.0
pandas==2.2.2
numpy==2.0.0
torch==2.3.1
torchaudio==2.3.1
transformers==4.42.3
pyaudio==0.2.14
tqdm==4.66.4
tensorflow==2.16.2
tf-keras==2.16.0
Wave==0.0.2
dotenv==0.0.5
